python/redis/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["manupanand <manupanandindianinvestor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
redis = "^5.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
